Dossier_review/src/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9

# HTTP
requests>=2.31.0
python-dotenv>=1.0.1

# Date parsing
dateparser>=1.2.0
python-dateutil>=2.9.0.post0

# PDF handling
pymupdf>=1.24.0

# Office formats
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.3.0
openpyxl>=3.1.2

# HTML parsing
beautifulsoup4>=4.12.3
lxml>=5.1.0

# OCR
pytesseract>=0.3.10
Pillow>=10.3.0

# Flowchart graph extraction (CV)
opencv-python-headless>=4.8,<4.11
numpy<2

# Grammar + Spelling
language-tool-python>=2.9.0
pyspellchecker>=0.8.1

# Docling
docling

# VLM (optional)
mlx-vlm
